Pushing in more changes. Check description please.
- Separated opportunity related functions away from class in `opportunity.py`

- Utilized functions in `utility.py` as well as `opportunity.py`

- Tested by calling `reset_processed_status()` and seeing if the functions still worked and it did. Made sure to update those opportunities as well so they don't get spewed out by the discord bot.
