+++ opportunity.py
@@ -1,49 +1,80 @@
 from dataclasses import dataclass
 from dotenv import load_dotenv
 import psycopg2
+from typing import List
 import os
+import utility as utils 
 
 load_dotenv()
 
-db_uri = os.getenv("DB_URI")
 table_name = os.getenv("DB_TABLE")
-
-
-def instantiates_db_connection():
-    """Returns the connection from the DB"""
-
-    return psycopg2.connect(db_uri)
-
 
 @dataclass
 class Opportunity:
     """Class for ..."""
+    # TODO
 
-    list_of_objects = [] # A list of objects holding all jobs (processed and not processed)
-    job_objects = {}  # A object populated with a job
 
-    def list_all_opportunities(self) -> "Opportunity":  # Return type of opportunity
-        """Lists all oppportunities in DB as well as returns them"""
+def list_all_opportunities() -> List[object]:  # Return type of opportunity
+    """Lists all oppportunities in DB as well as returns them"""
 
-        with instantiates_db_connection() as connection:
-            cursor = connection.cursor()
+    with utils.instantiates_db_connection() as connection:
+        cursor = connection.cursor()
 
-            cursor.execute(f"SELECT * FROM {table_name}")
+        cursor.execute(f"SELECT * FROM {table_name}")
 
-            rows = cursor.fetchall()
+        rows = cursor.fetchall()
 
-            for row in rows:
-                self.job_objects = {}
+        list_of_objects = []
 
-                self.job_objects["_company"] = row[0]
-                self.job_objects["_title"] = row[1]
-                self.job_objects["_location"] = row[2]
-                self.job_objects["_link"] = row[3]
-                self.job_objects["_processed"] = row[4]
+        for row in rows:
+            job_objects = {}
 
-                self.list_of_objects.append(self.job_objects)
+            job_objects["_company"] = row[0]
+            job_objects["_title"] = row[1]
+            job_objects["_location"] = row[2]
+            job_objects["_link"] = row[3]
+            job_objects["_processed"] = row[4]
 
-                # Uncomment to view all jobs
+            list_of_objects.append(job_objects)
+
+            # Uncomment to view all jobs
+
+            # _company, _title, _location, _link, _processed = row
+
+            # print("Company:", _company)
+            # print("Title:", _title)
+            # print("Location:", _location)
+            # print("Link:", _link)
+            # print("Processed:", _processed)
+            # print(" ")
+
+    return list_of_objects
+
+
+
+        
+def list_filtered_opportunities() -> List[object]:
+    """Returns a List[object] of job data that have a status of _processed = 0"""
+
+    with utils.instantiates_db_connection() as connection:
+        cursor = connection.cursor()
+
+        cursor.execute(f"SELECT * FROM {table_name} WHERE _processed = 0 LIMIT 20")
+        rows = cursor.fetchall()
+
+        list_of_objects = []
+
+        for row in rows:
+            job_objects = {}
+
+            job_objects["_company"] = row[0]
+            job_objects["_title"] = row[1]
+            job_objects["_location"] = row[2]
+            job_objects["_link"] = row[3]
+            job_objects["_processed"] = row[4]
+
+                # Uncomment to view up to 10 jobs that have not been posted
 
                 # _company, _title, _location, _link, _processed = row
 
@@ -54,6 +85,8 @@ class Opportunity:
                 # print("Processed:", _processed)
                 # print(" ")
 
-        return self
+            list_of_objects.append(job_objects)
+
+    return list_of_objects
